ai-email-assistant/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
beautifulsoup4
google-generativeai
markdown
